seperate resolver creation from routing
