+++ modules/server/src/mapping/createConnectionResolvers.ts
@@ -1,11 +1,8 @@
 import { IResolvers } from '@graphql-tools/utils';
 
-import { ConfigProperties, ExtendedConfigsInterface } from '@/config/types';
 import { GetServerSideFilterFn } from '@/utils/getDefaultServerSideFilter';
 
-import { applyAggregationMasking } from './masking';
-import resolveAggregations, { aggregationsToGraphql } from './resolveAggregations';
-import resolveHits from './resolveHits';
+import { createResolvers } from './resolvers';
 
 // TODO: tighten these types
 type CreateConnectionResolversArgs = {
@@ -28,38 +25,20 @@ const createConnectionResolvers: CreateConnectionResolversFn = ({
 	Parallel,
 	type,
 }) => {
-	const configs = async (parentObj, { fieldNames }: { fieldNames: string[] }) => {
-		return {
-			downloads: type.config?.[ConfigProperties.DOWNLOADS],
-			extended: fieldNames
-				? type.extendedFields.filter((extendedField: ExtendedConfigsInterface) =>
-						fieldNames.includes(extendedField.fieldName),
-				  )
-				: type.extendedFields,
-			...(createStateResolvers && {
-				facets: type.config?.[ConfigProperties.FACETS],
-				matchbox: type.config?.[ConfigProperties.MATCHBOX],
-				table: type.config?.[ConfigProperties.TABLE],
-			}),
-		};
-	};
-
-	const aggregationsQuery = resolveAggregations({ type, getServerSideFilter });
-	const aggregationsResolver = async (obj, args, context, info) => {
-		const aggs = await aggregationsQuery(obj, args, context, info);
-		return aggregationsToGraphql(aggs);
-	};
-
-	const defaultHitsResolver = resolveHits({ type, Parallel, getServerSideFilter });
-	const hitsResolver = enableDocumentHits
-		? defaultHitsResolver
-		: resolveHitsFromAggs(aggregationsQuery, dataMaskThreshold);
+	const { aggregations, hits, configs } = createResolvers({
+		createStateResolvers,
+		type,
+		Parallel,
+		getServerSideFilter,
+		dataMaskThreshold,
+		enableDocumentHits,
+	});
 
 	return {
 		[type.name]: {
-			aggregations: aggregationsResolver,
+			aggregations,
 			configs,
-			hits: hitsResolver,
+			hits,
 			// keeping this available for backwards compatibility, but hoping to remove it
 			// TODO: investigate its current usage and need. remove otherwise
 			// Update 2023-02: ENABLE_ADMIN prevents error comes up on facets.
@@ -72,36 +51,5 @@ const createConnectionResolvers: CreateConnectionResolversFn = ({
 		},
 	};
 };
-
-/**
- * Resolve hits from aggregations
- * If "aggregations" field is not in query, return 0
- *
- * @param aggregationsQuery - resolver ES query code for aggregations
- * @returns Returns a total count that is less than or equal to the actual total hits in the query.
- */
-const resolveHitsFromAggs =
-	(aggregationsQuery, dataMaskThreshold) => async (obj, args, context, info) => {
-		/*
-		 * Get "aggregations" field from full query if found
-		 * Popular gql parsing libs parse the "info" property which may not include full query based on schema
-		 */
-		const fileNameConnectionProperty = info.operation.selectionSet.selections[0];
-		const aggregationsSelectionSet = fileNameConnectionProperty.selectionSet.selections.find(
-			(selection) => selection.name.value === 'aggregations',
-		);
-
-		if (aggregationsSelectionSet) {
-			const modifiedInfo = { ...info, fieldNodes: [aggregationsSelectionSet] };
-			const aggregations = await aggregationsQuery(obj, info.variableValues, context, modifiedInfo);
-			const { hitsTotal: total } = applyAggregationMasking({
-				aggregations,
-				thresholdMin: dataMaskThreshold,
-			});
-			return { total };
-		} else {
-			return { total: 0 };
-		}
-	};
 
 export default createConnectionResolvers;
