+++ modules/server/src/mapping/resolvers.ts
@@ -1,0 +1,77 @@
+import { ConfigProperties, ExtendedConfigsInterface } from '@/config/types';
+import { applyAggregationMasking } from './masking';
+import resolveAggregations, { aggregationsToGraphql } from './resolveAggregations';
+import resolveHits from './resolveHits';
+
+/**
+ * Resolve hits from aggregations
+ * If "aggregations" field is not in query, return 0
+ *
+ * @param aggregationsQuery - resolver ES query code for aggregations
+ * @returns Returns a total count that is less than or equal to the actual total hits in the query.
+ */
+const resolveHitsFromAggs =
+	(aggregationsQuery, dataMaskThreshold) => async (obj, args, context, info) => {
+		/*
+		 * Get "aggregations" field from full query if found
+		 * Popular gql parsing libs parse the "info" property which may not include full query based on schema
+		 */
+		const fileNameConnectionProperty = info.operation.selectionSet.selections[0];
+		const aggregationsSelectionSet = fileNameConnectionProperty.selectionSet.selections.find(
+			(selection) => selection.name.value === 'aggregations',
+		);
+
+		if (aggregationsSelectionSet) {
+			const modifiedInfo = { ...info, fieldNodes: [aggregationsSelectionSet] };
+			const aggregations = await aggregationsQuery(obj, info.variableValues, context, modifiedInfo);
+			const { hitsTotal: total } = applyAggregationMasking({
+				aggregations,
+				thresholdMin: dataMaskThreshold,
+			});
+			return { total };
+		} else {
+			return { total: 0 };
+		}
+	};
+
+export const createResolvers = ({
+	createStateResolvers,
+	type,
+	Parallel,
+	getServerSideFilter,
+	dataMaskThreshold,
+	enableDocumentHits,
+}) => {
+	// configs
+	const configs = async (parentObj, { fieldNames }: { fieldNames: string[] }) => {
+		return {
+			downloads: type.config?.[ConfigProperties.DOWNLOADS],
+			extended: fieldNames
+				? type.extendedFields.filter((extendedField: ExtendedConfigsInterface) =>
+						fieldNames.includes(extendedField.fieldName),
+				  )
+				: type.extendedFields,
+			...(createStateResolvers && {
+				facets: type.config?.[ConfigProperties.FACETS],
+				matchbox: type.config?.[ConfigProperties.MATCHBOX],
+				table: type.config?.[ConfigProperties.TABLE],
+			}),
+		};
+	};
+
+	// aggregations
+	const aggregationsQuery = resolveAggregations({ type, getServerSideFilter });
+
+	const aggregations = async (obj, args, context, info) => {
+		const aggs = await aggregationsQuery(obj, args, context, info);
+		return aggregationsToGraphql(aggs);
+	};
+
+	// hits
+	const defaultHitsResolver = resolveHits({ type, Parallel, getServerSideFilter });
+	const hits = enableDocumentHits
+		? defaultHitsResolver
+		: resolveHitsFromAggs(aggregationsQuery, dataMaskThreshold);
+
+	return { hits, aggregations, configs };
+};
